pyproject.toml
[tool.poetry]
name = "notifications-sprint-1"
version = "0.1.0"
description = ""
authors = ["epacnhincev <alexei.epanchincev@yandex.ru>",
           "DBWtv <dima.bwc@gmail.com>",
           "jokcik <>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
psycopg2-binary = "^2.9.9"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
greenlet = "^3.1.0"
aiohttp = "^3.10.5"
gunicorn = "^23.0.0"
pika = "^1.3.2"
aio-pika = "^9.4.3"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pre-commit = "^3.7.1"

[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["src"]
skip_glob=[
    ".git",
    ".venv",
    ".vscode",
    "venv",
    "env",
    ".env",
    "*/migrations/*",
    "settings.py",
]
known_flask=["fastapi"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "FLASK",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"